Add pom.xml and update .gitignore
In this commit, a new 'pom.xml' file has been added to define the project with all its dependencies. This will set the project up for Maven build and deployment, and it assists in the project-specific configuration. The '.gitignore' file has been updated to exclude the 'target/' directory, which is generated during the build process.