The provided commit message does not match with the provided diffs. Based on the diffs, a correct commit message could be:
Allow LocalOperationExecutor to run in prod profile

The LocalOperationExecutor is now enabled to run under both the "local" and "prod" profiles. This change allows the same components to be used in both the development and production environments. In addition, an unnecessary line has been removed from pom.xml for clean and efficient coding.